lear-tg-bot/lear-tg-bot/requirements.txt
googletrans
pip==23.2.1
setuptools==65.5.0
python-telegram-bot
pydantic
openai
python-dotenv